requirements.txt
fastapi>=0.120.0
uvicorn>=0.38.0
sqlalchemy>=2.0.44
psycopg2-binary>=2.9.11
alembic>=1.17.1
bcrypt==4.1.3
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.5.0
python-multipart>=0.0.20
pydantic[email]>=2.12.0
pytest>=8.4.2
pytest-asyncio>=1.2.0
httpx>=0.28.1
